requirements.txt
qrcode
pillow
pywin32
